adding compile session questions
