+++ Quiz.Infrastructure/Quiz.Infrastructure.vcxproj
@@ -127,7 +127,8 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="server\QAModel.cpp" />
+    <ClCompile Include="server\CompileSessionQuestions.cpp" />
+    <ClCompile Include="server\Models\QAModel.cpp" />
     <ClCompile Include="src\Quiz.Infrastructure.cpp" />
     <ClCompile Include="server\server.cpp" />
     <ClCompile Include="client\startUp.cpp" />
